Revert "Google Drive connector refacto (#13174)"
This reverts commit 1d448d17c63563f73f70b1958f3e200a03f0ff00.
