--- connectors/src/connectors/google_drive/temporal/activities/common/types.ts
@@ -1,5 +1,0 @@
-export type LightGoogleDrive = {
-  id: string;
-  name: string;
-  isSharedDrive: boolean;
-};
