--- connectors/src/connectors/google_drive/temporal/activities/index.ts
@@ -1,11 +1,0 @@
-export { fixParentsConsistencyActivity } from "./fix_parents_consistency_activity";
-export { garbageCollector } from "./garbage_collector";
-export { garbageCollectorFinished } from "./garbage_collector_finished";
-export { getDrivesToSync } from "./get_drives_to_sync";
-export { getFoldersToSync } from "./get_folders_to_sync";
-export { incrementalSync } from "./incremental_sync";
-export { markFolderAsVisited } from "./mark_folder_as_visited";
-export { populateSyncTokens } from "./populate_sync_tokens";
-export { shouldGarbageCollect } from "./should_garbage_collect";
-export { syncFiles } from "./sync_files";
-export { upsertSharedWithMeFolder } from "./upsert_shared_with_me_folder";
