+++ core/bin/core_api.rs
@@ -4251,11 +4251,11 @@ fn main() {
         )
         .route(
             "/projects/:project_id/data_sources/:data_source_id/tables/:table_id/rows",
-            post(tables_rows_upsert), // yolo
+            post(tables_rows_upsert),
         )
         .route(
             "/projects/:project_id/data_sources/:data_source_id/tables/:table_id/csv",
-            post(tables_csv_upsert), // yolo
+            post(tables_csv_upsert),
         )
         .route(
             "/projects/:project_id/data_sources/:data_source_id/tables/:table_id/rows/:row_id",
